Remove circular dependency
Previous dependency loop:
- `utils` imports `Linter`
- `Linter` imports 'utils'

Now:
- `linter_factory` imports `Linter`
- `Linter` imports 'utils'
- `utils` does not import `Linter` any more
